[SE-2741] Clean up tmp folder if OCIM appserver provisioning fails
